user story 9 finished
you can modify the souvenir list by logging in as a administrator, go to manage stadiums, select the stadium you want to modify, then click the modify souvenirs buttion to do the operation.
modify a sourvenir: use the drop down menu to select which sourvenir to modify, then enter the info in the textBox, at the end click modify.
delete a sourvenir: use the drop down menu to select which sourvenir to delete, then click delete.
add a sourvenir: enter the info in the textBox, then click add.